gh-133532: Run GC fast cycles test in subprocess.
